classifiers/webapp/requirements.txt
nltk
scikit-learn==1.2.2
numpy==1.26.4
torch
torchvision
transformers
pillow
joblib
flask